Simplify groups list loading
