+++ app/views/access_token_groups/_error.html.erb
@@ -1,4 +1,4 @@
-<div id="group-select-wrapper">
+<div id="groups-select">
   <select name="feed[target_group]" id="feed_target_group" class="form-select" disabled>
     <option value="">Error loading groups</option>
   </select>
