+++ app/views/access_token_groups/index.turbo_stream.erb
@@ -1,15 +1,5 @@
-<%= turbo_stream.replace "group-select-wrapper" do %>
-  <div id="group-select-wrapper">
-    <select name="feed[target_group]" id="feed_target_group" class="form-select">
-      <option value="">Choose target group...</option>
-      <% @managed_groups.each do |group| %>
-        <option value="<%= group[:username] %>">
-          <%= group[:screen_name].present? ? "#{group[:screen_name]} (#{group[:username]})" : group[:username] %>
-          <%= "🔒" if group[:is_private] %>
-        </option>
-      <% end %>
-    </select>
-    <div class="form-text">Select the Freefeed group where posts will be published</div>
-    <div class="invalid-feedback">Please select a target group.</div>
+<%= turbo_stream.replace "groups-select" do %>
+  <div id="groups-select">
+    <%= render "feeds/groups_select", groups: @managed_groups, selected_group: nil, disabled: false %>
   </div>
 <% end %>
